Initiated new components
Basket is the parent component of the following components. It includes the preferences of the customer and the button to append the purchase list.

Products is a section to observe the quantities of the products that the user   wants to purchase.

AddToBasket is the button to finalize the preferences and appending the purchase list.